Create VM and Template methods

+++ src/main/java/controllers/VMTemplateController.java
@@ -1,6 +1,7 @@
 package controllers;
 
 import org.opennebula.client.ClientConfigurationException;
+import org.opennebula.client.template.Template;
 import service.onedb.GroupService;
 import service.onedb.UsersOneService;
 import service.onedb.VMTemplateService;
@@ -20,14 +21,24 @@ public class VMTemplateController {
     }
 
     public void createTemplate(String courseName, String userName, String imageName) throws ClientConfigurationException, IOException {
-        if (groupService.getGroupById(groupService.getGroupId(courseName))
-                .contains(usersOneService.getUserId(userName))) {
-            vmTemplateService.createVMTemplate(imageName).chown(usersOneService.getUserId(userName));
-        }
+//        if (groupService.getGroupById(groupService.getGroupId(courseName))
+//                .contains(usersOneService.getUserId(userName))) {
+            Template newTemplate= vmTemplateService.createVMTemplate(imageName);
+            newTemplate.chown(usersOneService.getUserId(userName));
+            newTemplate.chgrp(groupService.getGroupId(courseName));
     }
 
     public void removeTemplate(String userName, int templateId) throws ClientConfigurationException, IOException {
-        vmTemplateService.deleteTemplate(templateId);
+        if (vmTemplateService.getTemplateById(templateId).uid() == usersOneService.getUserId(userName)) {
+            vmTemplateService.deleteTemplate(templateId);
+        }
     }
+
+    public void changeTemplatePermissions(String userName, int templateId, int permissionCode) throws ClientConfigurationException {
+        if (vmTemplateService.getTemplateById(templateId).uid() == usersOneService.getUserId(userName)) {
+            vmTemplateService.changePermissions(templateId, permissionCode);
+        }
+    }
+
 }
